+++ src/features/library/utils/endLibraryStateActivity.ts
@@ -1,0 +1,15 @@
+import {isAndroid} from '../../../utils/android/isAndroid';
+import {LibraryStateNotificationBridge} from '../../../utils/android/libraryStateNotificationBridge';
+import {LibraryDynamicIslandBridge} from '../../../utils/ios/libraryDynamicIslandBridge';
+import {
+  deleteLibrarySeatStartTime,
+  deleteLibraryUsingStatus,
+} from '../storage/library';
+
+export const endLibraryStateActivity = () => {
+  if (isAndroid) LibraryStateNotificationBridge.end();
+  else LibraryDynamicIslandBridge.onEndActivity();
+
+  deleteLibrarySeatStartTime();
+  deleteLibraryUsingStatus();
+};
