+++ src/store/app/libraryStateActivity.ts
@@ -1,0 +1,7 @@
+import atomWithMMKV from '../../storage/atomWithMMKV';
+
+export const initIsLibraryStateActivity = true;
+export const libraryStateActivityAtom = atomWithMMKV<boolean>(
+  'isLibraryStateActivity',
+  initIsLibraryStateActivity,
+);
