+++ src/utils/android/isAndroid.ts
@@ -1,0 +1,3 @@
+import {Platform} from 'react-native';
+
+export const isAndroid = Platform.OS === 'android';
